updated trips migration to use decimals

+++ database/migrations/2023_05_19_223919_create_trips_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('car_id');
             $table->datetime('date');
-            $table->integer('miles');
+            $table->decimal('miles');
             $table->timestamps();
 
             $table->foreign('car_id')
